Do not reset virtualedit setting
If selection=exclusive and the end of the match is at the end of the
line, the cursor should move one column behind the actual match, so that
yanking works as expected. However, since we did reset the `ve` setting
right before the function returns, this would have also changed the
cursor position back to the end of the match, which means that the
character the cursor is at, won't be yanked (because of
selection=exclusive).

So do not reset `ve=` after the function returns. This has the
disadvantage, that the `ve` setting leaks into the users environment,
but at least it works as expected than.

ref: #19